airflow/requirements.txt
firebase-admin>=6.0.0
psycopg2-binary>=2.9.0
google-cloud-firestore>=2.0.0
cryptography>=3.4.0
requests>=2.28.0
